fix: increase FormatorId size with long

+++ WSRS-SWAFO/Models/ReportPending.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace WSRS_SWAFO.Models
 {
@@ -7,7 +6,7 @@ namespace WSRS_SWAFO.Models
     {
         [Key]
         public int Id { get; set; }
-        public int FormatorId { get; set; }
+        public long FormatorId { get; set; }
         public string Formator { get; set; } 
         public string Description { get; set; }
         public DateOnly ReportDate { get; set; }
